refactor: PathService 메서드 분리 및 Fare 불필요 생성자 삭제
