+++ src/main/java/wooteco/subway/exception/path/InvalidPathException.java
@@ -1,14 +1,13 @@
-package wooteco.subway.path.application;
+package wooteco.subway.exception.path;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class InvalidPathException extends RuntimeException {
-    public InvalidPathException() {
-    }
+    private static final String MESSAGE = "경로를 조회할 수 없습니다.";
 
-    public InvalidPathException(String message) {
-        super(message);
+    public InvalidPathException() {
+        super(MESSAGE);
     }
 }
