+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -18,9 +18,10 @@ import java.util.Objects;
 @Service
 @Transactional
 public class PathService {
-    private LineService lineService;
-    private StationService stationService;
-    private PathFinder pathFinder;
+
+    private final LineService lineService;
+    private final StationService stationService;
+    private final PathFinder pathFinder;
 
     public PathService(LineService lineService, StationService stationService, PathFinder pathFinder) {
         this.lineService = lineService;
@@ -29,22 +30,20 @@ public class PathService {
     }
 
     public PathResponse findPath(Long source, Long target, LoginMember loginMember) {
-        try {
-            List<Line> lines = lineService.findLines();
-            Station sourceStation = stationService.findStationById(source);
-            Station targetStation = stationService.findStationById(target);
-            SubwayPath subwayPath = pathFinder.findPath(lines, sourceStation, targetStation);
+        List<Line> lines = lineService.findLines();
+        Station sourceStation = stationService.findStationById(source);
+        Station targetStation = stationService.findStationById(target);
+        SubwayPath subwayPath = pathFinder.findPath(lines, sourceStation, targetStation);
+        return getPathResponse(loginMember, subwayPath);
+    }
 
-            List<Station> stations = subwayPath.getStations();
-            int distance = subwayPath.calculateDistance();
-            Fare fare = new Fare(distance, subwayPath.calculateLineFare());
-            if (Objects.isNull(loginMember)) {
-                return PathResponseAssembler.assemble(stations, distance, fare.calculateBasicFare());
-            }
-            return PathResponseAssembler.assemble(stations,distance, fare.calculateDiscountFare(loginMember.getAge()));
-
-        } catch (Exception e) {
-            throw new InvalidPathException();
+    private PathResponse getPathResponse(LoginMember loginMember, SubwayPath subwayPath) {
+        List<Station> stations = subwayPath.getStations();
+        int distance = subwayPath.calculateDistance();
+        Fare fare = new Fare(distance, subwayPath.calculateLineFare());
+        if (Objects.isNull(loginMember)) {
+            return PathResponseAssembler.assemble(stations, distance, fare.calculateBasicFare());
         }
+        return PathResponseAssembler.assemble(stations,distance, fare.calculateDiscountFare(loginMember.getAge()));
     }
 }
