+++ src/main/java/wooteco/subway/path/dto/PathResponseAssembler.java
@@ -1,6 +1,5 @@
 package wooteco.subway.path.dto;
 
-import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationResponse;
 
@@ -8,9 +7,13 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class PathResponseAssembler {
+
+    private PathResponseAssembler() {
+    }
+
     public static PathResponse assemble(List<Station> stations, int distance, int fare) {
         List<StationResponse> stationResponses = stations.stream()
-                .map(it -> StationResponse.of(it))
+                .map(StationResponse::of)
                 .collect(Collectors.toList());
         return new PathResponse(stationResponses, distance, fare);
     }
